+++ packages/plugin-infra/src/web3-state/Hub/NonFungibleClient.ts
@@ -18,9 +18,9 @@ import {
 } from '@masknet/web3-shared-base'
 import { EMPTY_LIST } from '@masknet/shared-base'
 import type { AuthorizationAPI, NonFungibleTokenAPI, TokenListAPI } from '@masknet/web3-providers'
-import { HubStateClient } from '../Hub'
+import { HubStateBaseClient } from '../Hub'
 
-export class HubStateNonFungibleClient<ChainId, SchemaType> extends HubStateClient<ChainId> {
+export class HubStateNonFungibleClient<ChainId, SchemaType> extends HubStateBaseClient<ChainId> {
     protected getProviders(
         initial?: HubOptions<ChainId>,
     ): Array<
